chore: store the operator profiler in the execution options
This moves the operator profiler from direct storage in the context to the
execution options, which are pointed to by the execution dependencies. They can
be updated by the interpreter and the results immediately seen by sub-execution
steps. This removes the need to copy the profiler into the context in the
tableFind/chain functions.