feat: make profiler available to tableFind and friends (#3264)
* feat: make profiler available to tableFind and friends

The interpreter now installs profilers into the execution dependencies. Doing
this directly from the interpreter is not possible due to the import cycle.

 flux -> interpreter -> execute -> compiler -> interpreter -> execute

This is the same cycle that trigger the lifting of execution dependencies into
their own module. This time more of the execution state must get added in and
the simple lift fix doesn't work.

Instead adding an interface to the interpreter module that is specifically for
setting execution options. The lang package can then implement this interface
and install the options.

So this patch also includes the movement of execution dependencies back to the
execute module and using this new interface to also set the now option in the
execution dependencies.

Tested by setting profilers and now in a flux script and verifying the values
in the execution dependencies afterwards.

This patch does not work with repl, as some additional changes are required
there to make profiling work at all.

* fix: improve interface/struct naming on profile-table-options branch

The execution options configuration interface is now called ExecOptsConfig. The
struct containing the options remains called ExecutionOptions.

* chore: store the operator profiler in the execution options

This moves the operator profiler from direct storage in the context to the
execution options, which are pointed to by the execution dependencies. They can
be updated by the interpreter and the results immediately seen by sub-execution
steps. This removes the need to copy the profiler into the context in the
tableFind/chain functions.

Closes #3204