feat: add series cardinality function to influxdb package
This adds a `cardinality()` function to the influxdb package to
determine the cardinality of a bucket with a filter for a specific
range. This is a source and can be chained to flux transformations.

To implement this, a dependency interface was created to act as an
influxdb provider. This dependency will cover reading and eventually
writing to influxdb. The interface allows us to define flux
transformations that interact with influxdb both in a remote way, such
as through http calls, and embedded within influxdb itself.

The interface is divided into a provider and a reader. The provider
contains a complete list of the different types of readers which is one
for each one. This provider is documented to be able to change and it is
advised to use the `UnimplementedProvider` as a base to ensure that all
of the methods continue to be implemented.

The readers are divided into one specific method for each method of
reading. If a provider returns a reader interface, it is expected that
the functionality is supported.

This allows us to have different implementations and also to allow an
implementation to make decisions around which type of reader it wants to
use. For example, a direct storage access implementation might fallback
to the http provider when a host is used.

The interface for directly reading through a normal read has been
stubbed out but hasn't been implemented. But, this will allow us to
consolidate the reading code to a single location instead of relying on
the planner and allow more of the flux code to live in the flux
repository instead of influxdb.