feat: add fluxinit package as an alternative to importing builtin (#3280)
The builtin package imports the standard library, then performs costly
initialization work in its init() function. If you need the flux library in
your program, you must perform the initialization step regardless of the path
of your program, which is not desirable if you only wish to print help or
perform some other task not related to executing flux.

This package just imports the standard library, but leaves the final
initialization step up to the user.