fix(types): format types in error message according to grammar
Previously the type error messages would format into the old style
grammar for type expressions. Now they format into a type expression
that is valid according to the Flux grammar.

This is essentially duplicate code as what is in the formatter, but the
formatter operates on AST nodes and here we have the actual MonoType
structs. It seems reasonable to implement for both types.

Fixes #3308