feat(formatter): use a tabwriter to vertically align tabstops (#3315)
This change adds the ability to vertically align records that appear in
an array. It uses an implementation of the tabwriter from BurntSushi
that implements and _elastic tabstops_ algorithm. Basically it ensures
that tabstops vertically align instead of treating them as fixed width
columns. See http://nickgravgaard.com/elastic-tabstops/, this is the
same method Go uses for its formatter.

I also bench marked this change to the formatter as I would expect it to
slow it down considerably. A benchmark on formatting the everything.flux
file run 14% slower with this change. That seems tolerable for this kind
of change but am open to suggestions to improve it.