chore: adjust libflux workspace to be idiomatic (#3353)
_Obligatory: apologies for using the term "idiomatic" here; I couldn't
find a better word._

When working with a rust workspace, it is usually in the context of
multiple connected crates. Our use of multiple crates is needed and
makes sense. However, the crates aren't laid out the way a normal crate
would be laid out, and the `src` in the workspace is usually not there.
This patch simplifies the rust workspace layout, removes some overrides
in various `Cargo.toml` files, and fixes the build accordingly.

Fixes #2342