feat: added rust scanner (#3362)
* feat: ported the scanner to rust

To build the rust scanner you need ragel 7.0.1 installed. Consult
Dockerfile_build for instructions on how to install this.

* feat: preparing to add the rust scanner as a separate module

Both scanners can exist simultaneously for some time and it should be an easy
fix if we need to revert.

* feat: use the rust scanner, can switch back by changing the scanner struct

* fix: moved the rust-scanner to the new loction in core/src

After rebasing on master the file names need to update.

* fix: moved all of the rust scanner to a sub-module

Using the rust scanner instead of c is now just a matter of referecing
rust::Scanner instead of Scanner.

* fix: don't need to clone the token defs, can include scanner::**

Also renamed scan2() -> scan() and included rust scanner in generated targets.

* chore: call the original scanner

Reverting to disabling the rust scanner. To enable it simply use
scanner::rust::Scanner from the parser instead of scanner::Scanner.

* fix: added some explanation around errors if ragel-7 install is wrong

* fix: added the generated rust scanner code and removed dependency from Makefile

Adding the generated code and removing the dependency from the Makefile will
make it such that the rust library can be compiled, but users won't yet need
ragel 7 installed.

* fix: don't check fmt/clippy of the generated rust scanner