fix(stdlib/strings): substring method now works on more indices (#3375)
The substring method in the strings package now works properly when the
start or end index isn't within the bounds of the existing string. It
truncates these indices to be the start or end of the string. This
reduces the need to check the length of the string to get a substring of
a certain size and makes it work more similar to how the javascript
substring operates.

This also fixes the case where both the start and the end index are
zero. This was erroneously creating a string of 1 character when it
should have been zero.