fix(lang): retrieve now option before running the flux script
This changes the `AstProgram` to retrieve the now option before it runs
the flux script. The reason is because the now option is placed into the
scope before the script is executed, but it can be reassigned to a
different value during execution. When we tried to retrieve the option
afterwards, it resulted in selecting the local now instead of the one
from the universe package.

We now retrieve the option ahead of time. If the option is redefined
with `option`, then the retrieved option will be modified in a mutable
way that we can access anyway and it avoids accidentally using a similar
variable with the same name.