fix(compiler): do not allow containers within a record to be null
When we retrieve a missing parameter from a record, we should not allow
it to return a null value if we know we are supposed to be returning one
of the container types (array, record, or dict). Those three are not
nullable so it is not valid for null to be returned. Instead, an error
is returned instead.