feat: add a pure flux test runner (#3388)
* feat: add a pure flux test runner

This patch adds the first pass at a pure flux test runner using the new
`testcase` statement. It ports the existing strings tests to `testcase`
tests to demonstrate the differences between the tests. The new tests
can be invoked with `go run ./cmd/flux test`.

Additionally, as the `TestCaseStatement` won't survive semantic checks,
the generation of existing tests must ignore the new-style `testcase`
tests entirely.

There are still a number of rough edges to be worked out, namely:

  - Lack of good output in the case of failures. Currently, the
expectation is that if a result is returned at all, it's the result of
`testing.diff`, and should be treated as a failure, but that result uses
`table.Stringify`. The test output in general could use some work.
  - Command line arguments to `flux test` are missing entirely, but
should be added for pathing and verbosity at the very least. It's likely
that, as other downstream projects implement their own runner, new flags
will be needed.

Fixes #3298

* fix: fix issue where compilation fails

* chore: address code review comments

* chore: update generated files

* chore: update doc comments

* chore: minor nit