refactor: clean up flux test runner
This patch refactors the new flux test runner to use structs to form a
cleaner organization and architecture for expansion into downstream
consumers. The intention is to layout out pieces that should be
composable. As such, the `TestRunner` and `TestReporter` are new in this
patch, though not in any sort of final form.

Additionally, cli arguments for the basic needs of the test runner were
introduced in this patch.