fix(lang): change extern parsing code to return a more descriptive error message (#3393)
The error parsing code returned an internal error from the rust library
and gave a cryptic "error getting json handle" which was specific to
the libflux and go interaction. This has changed the library to make the
parser an invalid error and make the error message specifically cite
that the extern json was what caused the error.