feat(tests): add ways to execute tests from tar and zip archives
The test command can now execute tests from tar or zip archives. It can
also execute tests from individual `.flux` files that are specified on
the command line.

This can be used for downloading a zip archive of the flux source code
from github and then executing the tests without unzipping the file.