chore: stdlib/universe: simplify difference implementation
This prepares for implementing https://github.com/influxdata/flux/issues/3400
by regularizing the implementation somewhat.

The `difference` type is now responsible only for returning
values and validity of those values (including dealing with `nonNegative`);
the `Process` method deals with removing the first row.

We rename the `difference.first` field to `difference.valid` (reversing
its sense in the process) because it's currently a little misleading,
because it might be true even when an update method is called
for a record that's not first.

Also reduce indentation somewhat by returning/continuing early
when appropriate.