feat: validate IP addresses from the dialer Control function
The dialer provides a function that gets called after DNS resolution, and
before the network connection is initiated. Use that to validate IP addresses.
This replaces URL-based validation where the IP address is resolved before the
post, and during any redirects. This method prevents attacks where the IP
address is changed between validation and connect.