fix: simple post validation fail needs to use the custom flux http client
In order for validation to take effect we must use the custom http client in
flux/http. It's not enough to set the validator. This is because validation is
now happening in the dialer control function and it requires we make a custom
client. Previously this was only necessary to make the redirect work.