feat: validate IP addresses from the dialer Control function (#3431)
* feat: validate IP addresses from the dialer Control function

The dialer provides a function that gets called after DNS resolution, and
before the network connection is initiated. Use that to validate IP addresses.
This replaces URL-based validation where the IP address is resolved before the
post, and during any redirects. This method prevents attacks where the IP
address is changed between validation and connect.

In order for validation to take effect we must use the custom http client in
flux/http. It's not enough to set the validator. This is because validation is
now happening in the dialer control function and it requires we make a custom
client. Previously this was only necessary to make the redirect validation work.