feat(stdlib/testing/expect): add testing/expect package for test expectations
The `testing/expect` package has been added to the standard library. It
contains functions to mark down when a testcase is expected to perform
some action within the engine and needs to assert that expectation is
fulfilled by the end of the test run.

The function is invoked at the beginning of the testcase and the
expectation is checked after the testcase has been executed.