feat: use FnvHasher for hash maps
The rust default hasher prioritizes safety, which often means memory
safety, but in the case of hashing, also means cryptographic safety.
This means that it's slower than it needs to be for non-cryptographic
application.

This patch uses `fnv::FnvHasher`, an external dependency, to provide
hashing capabilities. In our benchmarks, this makes scanning and parsing
operations 5-25% faster.