feat(cmd/test): add extends capability to testcase block
A testcase can now be extended using a relative path to another file.
When used, the file will be read from the filesystem and inserted into
the testcase in a wrapped context. Any testcase blocks within that
import will be wrapped in a function so they can be invoked from the
current testcase.

This has certain restrictions. Recursive extension of testcases isn't
allowed. Options are allowed, but this may cause a conflict if the
option is specified multiple times so this is discouraged for testcases
that will need to be imported from another testcase.