feat(csv): add support for parsing csv files without annotations
This change allows for the parsing of csv files without annotations.
The parsed data produces a single table where all data is of type
"string". All rows are considered to be part of the same group.

Additionally support is added to skip the UTF8 BOM and some parsing
logic around an extraLine was simplified to use a buffered csv reader.