feat(ast/astutil): create astutil package for ast utilities like formatting
The `ast/astutil` package can be used to format the text using the
libflux formatter. It accepts a file and produces a string using the
formatter.

The `ast/astutil` package depends on `libflux` and should only be
imported by packages that expect to have a `libflux` dependency.

Also, mapped go ast struct & rust ast struct with comments fields
added tests to check comments are getting preserved.

Co-authored-by: Sunil Kartikey <s.kartikey@gmail.com>