feat(fluxtest): add test modules to the testcase transform
The testcase transform now accepts a list of test modules with a
filesystem to access that module. The filesystem needs to follow the
`filesystem.Service` interface and attepting to extend a testcase will
lookup a testcase in the corresponding module and load the file
contents.

In order to extend a testcase, the testcase needs to be in a directory
where a `fluxtest.root` file is present and that file will have a name.
The file can then be extended with the following syntax:

    testcase b extends "flux/a_test" {
        a_test.testcase_name()
    }

This assumes that the `fluxtest.root` file has `"flux"` as the test
module name for this example.

The test command has now been updated so `-p` can take multiple paths,
archives, or files. If a file or directory is specified, the test
command will look in parent directories to find module roots. If
multiple module roots are found with the same name, an error will occur.