fix(cmd/test): find parent root should skip the current directory
If the path specified is a test root, the current code will try to add
that path twice which results in a duplicate test root error.

The parent root search is intended to find directories above the
filepath and not within the filepath. So this is now changed to find the
absolute path and then start with the parent directory of that path.

For the directory search, this means the parent directory. For the file
search, this means the directory that the file is located in.