chore: deny rust warnings again
While working on influxdata/flux-lsp#212, I noticed that we
accidentally allowed warnings somewhere along the line. When they are
specifically allowed, the compiler doesn't even tell us about them. This
patch re-enables the denial of warnings (and fixes the fallout) as well
as adds the strict policy to `flux` and `fluxdoc` crates.