chore: update rustup install command (#3590)
The current rustup install command uses SSL v3.0 by default in `curl`.
SSL v3 has been EOL'd for quite some time, and even TLS 1.2 was EOL'd
last year. It *appears* that the migrotion/update policy with regard to
the rustup.rs server is to fail occasionally on the older protocols,
more often and often until it fails for good.

This patch updates the rustup commands to specifically use tls 1.2 as
described on [rustup.rs](http://rustup.rs). This change has already been
made in idpe.

Additionally, I discovered `libflux/Dockerfile`, which had a manually
specified rust version like in the flux-build image. I updated that and
added a note to ensure that we don't get drift between them again. The
_right_ fix here might actually just be to either (a) delete
`libflux/Dockerfile` altogether, or have it be based on the image from
`./Dockerfile_build`. Both of those are outside the scope of this patch.