chore: add semgrep with ci and single initial rule
This patch adds support for [semgrep](https://semgrep.dev/). It adds a
CI step that runs the generic CI ruleset for semgrep (which did find a
couple issues that were fixed recently). Additionally, it adds a
warning-level semgrep for our custom rules, and adds a single new rule,
which can't be enforced until #3593 can be completed all the way
through.