chore: rename core crate to flux-core and remove stdlib re-export
There is already a rust standard crate called `core`, and this one
covers that one up.

The `pub use std::fmt` fixed a problem that came about because we named
our crate `core`, and a macro was trying to import `core::fmt`. Now that
the crate has been renamed, this workaround can be removed.

Fixes #3605