fix(execute): wait for the dispatcher to finish before finishing query
This modifies the execution engine to wait for the dispatcher to finish
before the query is finally closed. It also adds a cancel context to the
execution engine that gets threaded into the dispatcher so the
dispatcher can be interrupted. This gets attached to the default dataset
which will abort processing tables.