fix(execute): remove deadlock when an error happens while the dispatcher is stopping (#3723)
If the dispatcher had an error, such as a panic, while stopping, it
would attempt to set the error on the dispatcher. The `Stop()` method,
which could be triggered by a previous error, would take the lock and
then wait for the workers to finish.

But a worker could then try to set the error before it exited and
setting an error also attempted to take the lock. This created a
deadlock where stopping required waiting for the worker, but the worker
was waiting for stop to release the lock.

This modifies the dispatcher so that it does not hold the lock while it
is waiting and only holds it while using the shared state.