fix(execute): log when uneven columns are encountered when processing tables (#3744)
This produces a log message when a transformation is about to process a
column reader within a table that has uneven columns without aborting
execution.

This change only affects the transformation that receives the table. It
does not affect yields and it does not print the origin transformation
as it would have been difficult to get the logger threaded into the area
where the table itself is produced.

Instead, we wrap the table and validate each column reader as it is
being processed by the downstream.