fix(release): remove need for xcc.sh in release (#3760)
`xcc.sh` was a bash script to wrap various cross compiling targets and
select the compiler based on the environment. When the
`Dockerfile_build` image was simplified, `xcc.sh` was (accidentally)
removed, but the reference to its use in our goreleaser config was not
removed. This broke the release process.

This patch removes the need for `xcc.sh` in goreleaser, as we only
produce a linux amd64 binary now, and that's the default `CC` anyway. It
then also removes `xcc.sh` from the tree, as it's no longer needed.

Fixes #3757