perf(stdlib): optimize window and sort transformations (#3644)
* perf(stdlib): optimize sort transformation

This optimizes the sort transformation to perform a sort on each
individual buffer and then merge those buffers after all of the buffers
have been read. This is in contrast to constructing one huge table and
then using Go's sort from the standard library.

* perf(stdlib): optimize window transformation

This optimizes the window transformation to more efficiently organize
and construct tables for different groups. In particular, this
transformation will ensure that the input is sorted by time, constructs
the windows, and then determines the span of rows that go into each
window. This allows it to allocate the correct amount of memory and
avoid appending in random access patterns.

This also includes a planner optimization that inserts a sort by time
transformation in front of the window in certain circumstances.

When the window is preceded by group, keep, or drop, which are common
methods that regroup the data in multiple buffers, insert a sort by time
before the window.

This was found through experimenting to increase the speed of the window
to be faster than windowing without the sort. The sort helps standardize
the input into optimal buffer sizes and ensures that windows are
inserted to the minimum number of times that is possible. This speeds up
the overall runtime of window to make the additional time sorting or
buffering data as worth it. In particular, the group/keep/drop already
buffer the data so the sort doesn't do anything that hasn't been done
already.