feat(stdlib/pagerduty): add exclude parameter to pagerduty.dedupKey() (#3793)
Add an `exclude` parameter to the `pagerduty.dedupKey()` call. If
specified, this will ignore the given keys from the group key when
calculating the dedup key. If the columns don't exist, no error is
triggered.

The default for this value is `["_start", "_stop", "_level"]` which are
the typical group keys in from the monitor package that cause problems
for the dedup key.