fix(execute): allow work queue to be resized when work exceeds queue length
This modifies the dispatcher to make the work queue dynamically increase
when the amount of work exceeds the default size.

Previously, the work queue was implemented with a channel. Channels have
a static size and it was possible for a worker, which reads from the
work queue, to also attempt to add to the work queue and this resulted
in a deadlock.

This updates the work queue to be a dynamically sized ring to avoid a
deadlock when the amount of work exceeds the size of the work queue.