fix(semantic): properly copy record types with no extends parameter (#3888)
It seems that the code to copy a monotype did not copy the monotype
properly when the type was a record and it did not extend another
record. This causes compiler substitution to mark itself as extending
`t0` when the original was a fully qualified type with no record
extension.

This may have affected the `map()` call which could possibly heavily
rely on this code or it could have affected some other part of the
compiler infrastructure.

This was found while testing some code for an experiment so I don't have
a reproducer for how this would affect a live system.