fix(libflux): restore the previous libflux include paths to use pkg-config
The libflux package was modified to manually specify `CFLAGS` for the
include path. This worked for local builds where the files were located
in the source directory, but didn't work for builds that used flux as a
dependency like influxdb.

This restores the old behavior that uses the `/include` directory as the
base of the include path and includes `"influxdata/flux.h"` instead of
directly including `"flux.h"`. This should restore the previous
behavior.

It is up to `pkg-config` to produce a usable include path.