chore(libflux): support rust 1.54 (#3905)
This patch enables libflux to build on rust 1.54. The detection of
`dead_code` has gotten better, and discovered a struct property that we
set but never read. This patch removes it.

This patch specifically doesn't update the rust version in CI, as I
think these updates can and should be independent of each other.