feat(array): create flux array package (#3897)
The flux array package has been created to be the location where all
flux array implementations are kept.

The majority of the array implementations are type aliases to the
equivalent arrow array implementation and the interfaces and builders
have also been aliased or forwarded to arrow. Others, like string, are
wrappers around arrow implementations.

All of the locations that previously referenced the arrow array
implementations have now been changed to flux array implementations.

The intention behind this is so that we can control the naming and
implementations of the column representations while not having to modify
all the locations where the code is used. This is most useful with
string where we are creating our own more optimized implementation, but
helps with consistency across the different packages.