feat(execute): expand message interface with message lifetime controls (#3913)
The `execute.Message` interface has been expanded with `Ack()` and
`Dup()` methods.

The `Ack()` method is meant to acknowledge that the message has been
received by a `Transport` and the contents of the message can be
released.

The `Dup()` method is meant to copy the inner contents of a message into
a new message. Both messages must be acked separately. The
implementation of `Dup()` may involve incrementing a reference count
instead of a full deep copy.

These lifetimes will allow messages to contain data that has to be
released.