feat(array): add optimized repeat function for arrow arrays
This adds a `Repeat()` method to the `array` package that mimics the
pattern in from the `flux/arrow` package.

This `Repeat()` differs slightly in that it takes an interface as its
input instead of a `values.Value` and it differs in that it composes an
optimized version of string without repeated iterations over the string
builder.

The `arrow.Repeat()` function has been modified to unwrap the value and
make use of the `array.Repeat()` function instead of its own
implementation.