feat(execute): add two additional message types (#3916)
This adds two additional message types: `ProcessTableChunk` and `FlushKey`.

The `ProcessTableChunk` message represents that a chunk of a table is
ready to be processed by the transformation. It contains a reference to
the table chunk which is an arrow table buffer.

The `FlushKey` message represents that a group key has been finished and
is ready to flush any data stored in an associated dataset.