feat(execute): simplify the transport interface and add a transformation adapter
The `Transport` interface is now just `ProcessMessage` which can be
implemented directly by a transformation rather than using the
`Transformation` interface.

A transformation adapter has been created to wrap the older
`Transformation` into a `Transport` to support the new message types and
adapt them for the `Transformation` interface.

`consecutiveTransport` has been updated to implement both `Transport`
and `Transformation` to allow it to be used with either a
`Transformation` or a `Transport` and to forward the message to the
appropriate methods.