feat(execute): add transport-aware dataset (#3934)
The transport-aware dataset, or `TransportDataset`, can be used to send
messages directly to downstream transports. The `TransportDataset`
implements the `Dataset` interface and can be used in any place that
requires a Dataset.

If the downstream transformation implements the `Transport` interface,
the messages will be sent directly. If the downstream transformation
only implements the transformation interface, the transformation will be
wrapped in the transport adapter and that will be used instead.

If there are multiple downstream nodes, messages will be automatically
duplicated.