fix(execute): transformation adapter does not error when receiving a flush key for a table that is not present (#3950)
The transformation adapter would respond to the flush key message by
attempting to create a table for the key. If there was no table builder
associated with the key, it would cause an error instead of just
ignoring the message.

This updates the transformation adapter to ignore flush key messages for
keys that don't correspond to an existing table builder.