fix(stdlib): change die error code to invalid (#3957)
Previously the error code returned from the "die" function was
"Internal". Internal errors represent a bad state or system failure,
since a user can call the die function it is not indicative of an
internal error rather a user error as the system is behaving
appropriately when it returns the error. The change make the error code
"Invalid" instead of "Internal".