feat(execute): modify aggregate transformation to implement transport
This modifies the existing aggregate transformation to now be
`SimpleAggregateTransformation` and creates another
`AggregateTransformation` that is more general.

Both of these implement transport directly so they have support for
table chunks and flush key messages. This should work directly with any
transformations that utilize that interface.

The simple aggregate transformation is unmodified in how it is used.
There is an aggregate config along with an aggregate interface that can
be used to create the aggregate state. This will continue to work for
simple aggregates like min, max, and last.

The new aggregate transformation implements a more general interface
with an `Aggregate` function, which takes table chunks and
creates/manipulates state for a group key, and a `Compute` function
which transforms that state into a table. This can be used with some
more complex aggregates like `distinct()` and `mode()` while also being
useful for some functions that keep state and can only construct
themselves after receiving a flush key message, like `sort()`.

This change does not affect any of the selector implementations.