fix(stdlib): limit now resets the offset after processing a partial buffer
The `limit()` function had two code paths. One path would decrement the
offset if the offset was greater than the length of the buffer so it
could track when it should stop discarding data. The second path would
process a partial buffer when the offset was less than the length of the
buffer.

Unfortunately, the second case had an error where it would not reset the
offset to zero after processing a partial buffer. It was supposed to
reset it to zero so that future buffers would be processed starting from
index zero.