fix(stdlib): panic unknown type in query with difference function
Issue -
The issue is with the columns type which is use to compute the
difference. If the col is anything else then int, float or time the
flux.ColType stayed unassigned causing builder.AddCol to fail.

Fix -
The fix is to add default case for the ColType, generating a user error