feat: embed compiled stdlib instead of compiling at runtime (#3623)
This adds a `go generate` step to write compiled semantic packages for
the standard library that can then be loaded by the go runtime to
execute the packages.

The compiled package has all non-test files concatenated together and
then analyzed together. The analyzed semantic graph is then marshaled as
a flatbuffer, gzipped, and written as an `.fc` file which stands for
"flux compiled".

These files are then embedded with the new `embed` package for go 1.16.
At runtime, these files will be accessed, deserialized into the
in-memory structure, and then used for execution instead of analyzing
them on startup. This reduces startup time drastically.

Each package is written to a separate file. This is done so that changes
to a file will affect only the files that were touched. While this is
the intention, this may not actually happen because the rust code's
fresher is dependent on the imported packages. This can likely be fixed
later so that only the modified file changes.