revert: "feat: embed compiled stdlib instead of compiling at runtime (#3623)"
This reverts commit 9b9575705d78fe818e2d51b3318c2dd32f483e35.

Because of the way that type inference uses the fresher when analyzing
files, a change to one package would cause any packages that were
lexigraphically lower in the chain to also be rewritten because the
tvars were now different.

As this causes a rewrite of most files when changing flux code, this is
being reverted until type inference is modified to use a normalized
fresher for each file rather than sharing the one from the bootstrap
process.