feat(record): adds any value to record package (#4041)
Prior to this change it was not possible to use a default value that had
a record type because default values constrain the type of the function
parameter. See #3461 for a long term fix. In the medium term a valid
solution is to introduce a polymorphic record value that can be used in
these cases. This change adds record.any which is an empty record
but whose type is unbounded allowing for polymorphic values.