feat(plan): allow the query concurrency to be set to the number of nodes in the graph
This adds a feature flag to allow the concurrency quota to be equal to
the number of transformation nodes in the graph instead of equal to the
number of roots. This functionality gets enabled when the
`queryConcurrencyLimit` is set to a value higher than zero. This number
limits the maximum concurrency too.