fix(stdlib): events.duration() will properly handle multiple buffers
The `events.duration()` function in `contrib/tomhollingworth/events`
would always use the stop column or time to compare the last value in
each buffer rather than just the final row in the final buffer.

This fixes it so that only the final row will be compared with the stop
time instead of the final time in each buffer.