fix(stdlib): optimized derivative attempted to replicate a bug that didn't exist
The optimized derivative attempted to replicate a bug with the existing
derivative where it wouldn't error if the column was not a compatible
type with derivative. This is because I saw an issue detailing this and
I did not see where the check was.

A test was created to verify the bad behavior with a note to fix it in
the future with the issue.

After seeing this test fail when used with the older derivative, it came
to my attention that the bug had already been fixed and so the newer
derivative didn't have to replicate the bug.

This removes the test and adds the necessary code to mark the condition
with the same error.