fix: make builds reproducible again by ordering package members in docs (#4071)
The individual members for a package were serialized from a hash map in
an unpredictable order. This caused the `libflux.a` file to come out
differently on subsequent build runs.

This changes the `PackageDoc` structure to use a `BTreeMap` instead of a
`HashMap` to ensure the members are serialized consistently.