feat(libflux): adds Analyzer API to libflux
Prior to this change there were several entry points into the type
inference code. It was not clear which one was best or why you would use
one over the other. Now there is a single entry into the inference logic
via the Analyzer type.

Additionally the Analyzer has been implemented with a generic Importer
allowing for either analysis against embedded stdlib data or against a
file system containing the compiled types.

An implementation of the Importer against the file system was
implemented along with a fluxc command that dumps compiled stdlib into
a directory structure. This structure is similar to how the stdlib is
embedded into Go. In fact the explicit goal is to eventually merge those
two systems into a single system, but that work is left for a later PR.
These new files are also called .fc files and instead of only
containing the semantic graph also contain the poly type for the
package. This PR however does not encode the semantic graph into the new
.fc files as that is not yet implemented.

In summary these are the changes in the PR:

* New Analyzer type
* Removed lots of old public functions that did analysis in different ways
* Add implementation of Importer using the FileSystem
* Updated the bootstrap module to both read source code for the stdlib
* and output compiled stdlib files
* Adds Compiled flatbuffer type for the content stored in compiled .fc
* files. (Note this does not change the Go embed process).
* Updated serialization of semantic graph to be into a partial builder
* instead of always assuming nothing else will be serialized alongside.
* Add a fluxc command that dumps compiled stdlib into files using the
* Add a fluxdoc command that can generate docs json dumps from Flux
  source code
* bootstrap module.
* Analyzer has explicit error enum to communicate about the kinds of analysis error explicitly